refactor(dashboard): migrate medium panels — settings / memory / skills / mcp
Stage 4 PR 4.2 from #28. Four panels with bigger surfaces (forms,
detail views, mutation flows) move out of dashboard/app.js:

- panels/settings.ts (153 LoC) — config form (api key, base url,
  preset, effort, search, runtime info)
- panels/memory.ts (172 LoC) — REASONIX.md + global / project-mem
  list + textarea editor
- panels/skills.ts (185 LoC) — list + create + edit + delete
  across project / global / builtin scopes
- panels/mcp.ts (206 LoC) — bridged-server table, persisted-spec
  CRUD, server detail view (tools / resources / prompts)

dashboard/app.js loses 656 LoC. Bundle still ~122 KB; CDN
imports stay external. Typecheck + 1682 tests pass.

Sessions panel deferred — it depends on ChatMessage which lives
with the Chat panel migration (Stage 4 PR 4.6).
