+++ dashboard/app.js
@@ -26,9 +26,13 @@ import {
 } from "./src/lib/markdown";
 import { usePoll } from "./src/lib/use-poll";
 import { HooksPanel } from "./src/panels/hooks";
+import { McpPanel } from "./src/panels/mcp";
+import { MemoryPanel } from "./src/panels/memory";
 import { OverviewPanel } from "./src/panels/overview";
 import { PermissionsPanel } from "./src/panels/permissions";
 import { PlansPanel } from "./src/panels/plans";
+import { SettingsPanel } from "./src/panels/settings";
+import { SkillsPanel } from "./src/panels/skills";
 import { ToolsPanel } from "./src/panels/tools";
 
 const html = htm.bind(h);
@@ -1788,480 +1792,6 @@ function UsageWithChart() {
   `;
 }
 
-// ---------- Settings ----------
-
-function SettingsPanel() {
-  const [data, setData] = useState(null);
-  const [error, setError] = useState(null);
-  const [saving, setSaving] = useState(false);
-  const [saved, setSaved] = useState(null);
-  const [draft, setDraft] = useState({});
-
-  const load = useCallback(async () => {
-    try {
-      const r = await api("/settings");
-      setData(r);
-      setDraft({});
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-  useEffect(() => {
-    load();
-  }, [load]);
-
-  const save = useCallback(
-    async (fields) => {
-      setSaving(true);
-      setError(null);
-      try {
-        await api("/settings", { method: "POST", body: fields });
-        await load();
-        setSaved(`saved: ${Object.keys(fields).join(", ")}`);
-        setTimeout(() => setSaved(null), 3000);
-      } catch (err) {
-        setError(err.message);
-      } finally {
-        setSaving(false);
-      }
-    },
-    [load],
-  );
-
-  if (!data && !error) return html`<div class="boot">loading settings…</div>`;
-  if (error && !data) return html`<div class="notice err">${error}</div>`;
-  const v = data;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Settings</h2>
-        <span class="panel-subtitle">~/.reasonix/config.json · most fields apply next session</span>
-      </div>
-      ${saved ? html`<div class="notice">${saved}</div>` : null}
-      ${error ? html`<div class="notice err">${error}</div>` : null}
-
-      <div class="section-title">DeepSeek API</div>
-      <div class="card">
-        <div class="row">
-          <span class="card-title" style="margin: 0;">API key</span>
-          <code style="margin-left: auto;">${v.apiKey ?? "(not set)"}</code>
-        </div>
-        <div class="row" style="margin-top: 8px;">
-          <input
-            type="password"
-            placeholder="paste a fresh sk-… token to replace"
-            value=${draft.apiKey ?? ""}
-            onInput=${(e) => setDraft({ ...draft, apiKey: e.target.value })}
-          />
-          <button
-            class="primary"
-            disabled=${saving || !(draft.apiKey ?? "").trim()}
-            onClick=${() => save({ apiKey: draft.apiKey })}
-          >Save key</button>
-        </div>
-        <div class="row" style="margin-top: 12px;">
-          <span class="card-title" style="margin: 0;">Base URL</span>
-          <input
-            type="text"
-            value=${draft.baseUrl ?? v.baseUrl ?? ""}
-            placeholder="https://api.deepseek.com (default)"
-            onInput=${(e) => setDraft({ ...draft, baseUrl: e.target.value })}
-          />
-          <button
-            disabled=${saving || (draft.baseUrl ?? v.baseUrl ?? "") === (v.baseUrl ?? "")}
-            onClick=${() => save({ baseUrl: draft.baseUrl })}
-          >Save</button>
-        </div>
-      </div>
-
-      <div class="section-title">Defaults</div>
-      <div class="card">
-        <div class="row">
-          <span class="card-title" style="margin: 0; flex: 0 0 110px;">Preset</span>
-          <select
-            value=${
-              // Unknown values (legacy fast/smart/max, or anything
-              // hand-edited into config.json) display as `auto`.
-              ["auto", "flash", "pro"].includes(v.preset) ? v.preset : "auto"
-            }
-            onChange=${(e) => save({ preset: e.target.value })}
-            disabled=${saving}
-          >
-            <option value="auto">auto — flash → pro on hard turns (default)</option>
-            <option value="flash">flash — always flash, no auto-escalate</option>
-            <option value="pro">pro — always pro</option>
-          </select>
-          <span class="muted" style="margin-left: auto; font-size: 12px;">applies next turn</span>
-        </div>
-        <div class="row" style="margin-top: 12px;">
-          <span class="card-title" style="margin: 0; flex: 0 0 110px;">Effort</span>
-          <select
-            value=${v.reasoningEffort}
-            onChange=${(e) => save({ reasoningEffort: e.target.value })}
-            disabled=${saving}
-          >
-            <option value="max">max (default — best)</option>
-            <option value="high">high (cheaper / faster)</option>
-          </select>
-          <span class="muted" style="margin-left: auto; font-size: 12px;">applies next turn</span>
-        </div>
-        <div class="row" style="margin-top: 12px;">
-          <span class="card-title" style="margin: 0; flex: 0 0 110px;">Web search</span>
-          <button
-            class=${v.search ? "primary" : ""}
-            onClick=${() => save({ search: !v.search })}
-            disabled=${saving}
-          >${v.search ? "ON" : "off"}</button>
-          <span class="muted" style="margin-left: auto; font-size: 12px;">web_fetch + web_search tools</span>
-        </div>
-      </div>
-
-      <div class="section-title">Runtime</div>
-      <div class="card">
-        <div class="row">
-          <span class="card-title" style="margin: 0; flex: 0 0 110px;">Active model</span>
-          <code>${v.model ?? "—"}</code>
-        </div>
-        <div class="row" style="margin-top: 8px;">
-          <span class="card-title" style="margin: 0; flex: 0 0 110px;">Edit mode</span>
-          <code>${v.editMode}</code>
-          <span class="muted" style="margin-left: auto; font-size: 12px;">switch from the Chat tab header</span>
-        </div>
-      </div>
-    </div>
-  `;
-}
-
-// ---------- Hooks ----------
-
-// ---------- Memory ----------
-
-function MemoryPanel() {
-  const [tree, setTree] = useState(null);
-  const [error, setError] = useState(null);
-  const [open, setOpen] = useState(null); // { scope, name } | null
-  const [body, setBody] = useState("");
-  const [busy, setBusy] = useState(false);
-  const [info, setInfo] = useState(null);
-
-  const load = useCallback(async () => {
-    try {
-      const r = await api("/memory");
-      setTree(r);
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-  useEffect(() => {
-    load();
-  }, [load]);
-
-  const openFile = useCallback(async (scope, name) => {
-    setOpen({ scope, name });
-    setBusy(true);
-    try {
-      const path =
-        scope === "project" ? "/memory/project" : `/memory/${scope}/${encodeURIComponent(name)}`;
-      const r = await api(path);
-      setBody(r.body);
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, []);
-
-  const save = useCallback(async () => {
-    if (!open) return;
-    setBusy(true);
-    setError(null);
-    try {
-      const path =
-        open.scope === "project"
-          ? "/memory/project"
-          : `/memory/${open.scope}/${encodeURIComponent(open.name)}`;
-      await api(path, { method: "POST", body: { body } });
-      setInfo(`saved ${open.scope}${open.name ? `/${open.name}` : ""}`);
-      setTimeout(() => setInfo(null), 3000);
-      await load();
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, [open, body, load]);
-
-  if (!tree && !error) return html`<div class="boot">loading memory…</div>`;
-  if (error && !tree) return html`<div class="notice err">${error}</div>`;
-
-  if (open) {
-    return html`
-      <div>
-        <div class="panel-header">
-          <h2 class="panel-title">Memory · ${open.scope}${open.name ? `/${open.name}` : ""}</h2>
-          <button onClick=${() => setOpen(null)} style="margin-left: auto;">← back</button>
-        </div>
-        ${info ? html`<div class="notice">${info}</div>` : null}
-        ${error ? html`<div class="notice err">${error}</div>` : null}
-        <textarea
-          style="width: 100%; height: 480px; font-family: var(--mono); font-size: 13px; background: var(--bg-2); color: var(--fg-0); border: 1px solid var(--border); border-radius: var(--radius-md); padding: 12px;"
-          value=${body}
-          onInput=${(e) => setBody(e.target.value)}
-          disabled=${busy}
-        ></textarea>
-        <div class="row" style="margin-top: 8px;">
-          <button class="primary" disabled=${busy} onClick=${save}>Save</button>
-          <span class="muted" style="font-size: 12px;">${body.length.toLocaleString()} chars · re-applied on next /new or session restart</span>
-        </div>
-      </div>
-    `;
-  }
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Memory</h2>
-        <span class="panel-subtitle">REASONIX.md (committable) + private notes (~/.reasonix/memory)</span>
-      </div>
-      <div class="section-title">Project — REASONIX.md</div>
-      ${
-        tree.project.path
-          ? html`
-          <div class="card row" style="cursor: pointer;" onClick=${() => openFile("project")}>
-            <span><code>${tree.project.path}</code></span>
-            <span class="pill ${tree.project.exists ? "pill-ok" : "pill-dim"}" style="margin-left: auto;">
-              ${tree.project.exists ? "exists" : "create"}
-            </span>
-          </div>
-        `
-          : html`<div class="empty">No active project.</div>`
-      }
-
-      <div class="section-title">Global — ~/.reasonix/memory/global</div>
-      ${
-        tree.global.files.length === 0
-          ? html`<div class="empty">No global memory files yet.</div>`
-          : html`
-          <table>
-            <thead><tr><th>name</th><th class="numeric">size</th><th class="numeric">touched</th></tr></thead>
-            <tbody>
-              ${tree.global.files.map(
-                (f) => html`
-                <tr key=${f.name} style="cursor: pointer;" onClick=${() => openFile("global", f.name)}>
-                  <td><code>${f.name}</code></td>
-                  <td class="numeric">${fmtBytes(f.size)}</td>
-                  <td class="numeric muted">${fmtRelativeTime(f.mtime)}</td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-
-      ${
-        tree.projectMem.path
-          ? html`
-          <div class="section-title">Project private — ~/.reasonix/memory/&lt;hash&gt;</div>
-          ${
-            tree.projectMem.files.length === 0
-              ? html`<div class="empty">No project-private memory yet.</div>`
-              : html`
-              <table>
-                <thead><tr><th>name</th><th class="numeric">size</th><th class="numeric">touched</th></tr></thead>
-                <tbody>
-                  ${tree.projectMem.files.map(
-                    (f) => html`
-                    <tr key=${f.name} style="cursor: pointer;" onClick=${() => openFile("project-mem", f.name)}>
-                      <td><code>${f.name}</code></td>
-                      <td class="numeric">${fmtBytes(f.size)}</td>
-                      <td class="numeric muted">${fmtRelativeTime(f.mtime)}</td>
-                    </tr>
-                  `,
-                  )}
-                </tbody>
-              </table>
-            `
-          }
-        `
-          : null
-      }
-    </div>
-  `;
-}
-
-// ---------- Skills ----------
-
-function SkillsPanel() {
-  const [data, setData] = useState(null);
-  const [error, setError] = useState(null);
-  const [open, setOpen] = useState(null);
-  const [body, setBody] = useState("");
-  const [busy, setBusy] = useState(false);
-  const [info, setInfo] = useState(null);
-  const [newName, setNewName] = useState("");
-  const [newScope, setNewScope] = useState("global");
-
-  const load = useCallback(async () => {
-    try {
-      setData(await api("/skills"));
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-  useEffect(() => {
-    load();
-  }, [load]);
-
-  const openSkill = useCallback(async (scope, name) => {
-    setOpen({ scope, name });
-    setBusy(true);
-    try {
-      const r = await api(`/skills/${scope}/${encodeURIComponent(name)}`);
-      setBody(r.body);
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, []);
-
-  const save = useCallback(async () => {
-    if (!open) return;
-    setBusy(true);
-    try {
-      await api(`/skills/${open.scope}/${encodeURIComponent(open.name)}`, {
-        method: "POST",
-        body: { body },
-      });
-      setInfo(`saved ${open.scope}/${open.name}`);
-      setTimeout(() => setInfo(null), 3000);
-      await load();
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, [open, body, load]);
-
-  const remove = useCallback(async () => {
-    if (!open) return;
-    if (!confirm(`Delete skill ${open.scope}/${open.name}?`)) return;
-    setBusy(true);
-    try {
-      await api(`/skills/${open.scope}/${encodeURIComponent(open.name)}`, { method: "DELETE" });
-      setOpen(null);
-      await load();
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, [open, load]);
-
-  const create = useCallback(async () => {
-    if (!newName.trim()) return;
-    setBusy(true);
-    const stub = `---\nname: ${newName.trim()}\ndescription: TODO — one-line description that helps the model match this skill\n---\n\n# ${newName.trim()}\n\n`;
-    try {
-      await api(`/skills/${newScope}/${encodeURIComponent(newName.trim())}`, {
-        method: "POST",
-        body: { body: stub },
-      });
-      setNewName("");
-      await load();
-      openSkill(newScope, newName.trim());
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, [newName, newScope, load, openSkill]);
-
-  if (!data && !error) return html`<div class="boot">loading skills…</div>`;
-  if (error && !data) return html`<div class="notice err">${error}</div>`;
-
-  if (open) {
-    return html`
-      <div>
-        <div class="panel-header">
-          <h2 class="panel-title">Skill · ${open.scope}/${open.name}</h2>
-          <button onClick=${() => setOpen(null)} style="margin-left: auto;">← back</button>
-        </div>
-        ${info ? html`<div class="notice">${info}</div>` : null}
-        ${error ? html`<div class="notice err">${error}</div>` : null}
-        <textarea
-          style="width: 100%; height: 520px; font-family: var(--mono); font-size: 13px; background: var(--bg-2); color: var(--fg-0); border: 1px solid var(--border); border-radius: var(--radius-md); padding: 12px;"
-          value=${body}
-          onInput=${(e) => setBody(e.target.value)}
-          disabled=${busy}
-        ></textarea>
-        <div class="row" style="margin-top: 8px;">
-          <button class="primary" disabled=${busy} onClick=${save}>Save</button>
-          <button class="danger" disabled=${busy} onClick=${remove}>Delete</button>
-          <span class="muted" style="font-size: 12px; margin-left: auto;">re-loaded on next /new or session restart</span>
-        </div>
-      </div>
-    `;
-  }
-
-  const renderList = (label, items, scope) => html`
-    <div class="section-title">${label} (${items.length})</div>
-    ${
-      items.length === 0
-        ? html`<div class="empty">none</div>`
-        : html`
-        <table>
-          <thead><tr><th>name</th><th>description</th><th></th></tr></thead>
-          <tbody>
-            ${items.map(
-              (s) => html`
-              <tr key=${s.name} style="cursor: ${scope === "builtin" ? "default" : "pointer"};" onClick=${() => scope !== "builtin" && openSkill(scope, s.name)}>
-                <td><code>${s.name}</code></td>
-                <td>${s.description ?? html`<span class="muted">(no description)</span>`}</td>
-                <td>${scope === "builtin" ? html`<span class="pill pill-dim">builtin</span>` : null}</td>
-              </tr>
-            `,
-            )}
-          </tbody>
-        </table>
-      `
-    }
-  `;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Skills</h2>
-        <span class="panel-subtitle">click to edit · creates land in next /new</span>
-      </div>
-      ${error ? html`<div class="notice err">${error}</div>` : null}
-
-      <div class="section-title">Create new</div>
-      <div class="card row">
-        <select value=${newScope} onChange=${(e) => setNewScope(e.target.value)}>
-          <option value="global">global</option>
-          ${data.paths.project ? html`<option value="project">project</option>` : null}
-        </select>
-        <input
-          type="text"
-          placeholder="skill-name"
-          value=${newName}
-          onInput=${(e) => setNewName(e.target.value)}
-        />
-        <button class="primary" disabled=${busy || !newName.trim()} onClick=${create}>Create</button>
-      </div>
-
-      ${renderList("Project", data.project, "project")}
-      ${renderList("Global", data.global, "global")}
-      ${renderList("Builtin (read-only)", data.builtin, "builtin")}
-    </div>
-  `;
-}
-
-// ---------- MCP ----------
-
 // ---------- Semantic index ----------
 
 function SemanticPanel() {
@@ -2801,188 +2331,6 @@ function SkipBucketsView({ buckets }) {
     .filter(([k]) => (buckets[k] || 0) > 0)
     .map(([k, label]) => `${label}: ${buckets[k]}`);
   return html`<div><span class="kv-key">skipped</span>${total} files <span class="muted">(${parts.join(", ")})</span></div>`;
-}
-
-function McpPanel() {
-  const [data, setData] = useState(null);
-  const [specs, setSpecs] = useState(null);
-  const [error, setError] = useState(null);
-  const [info, setInfo] = useState(null);
-  const [newSpec, setNewSpec] = useState("");
-  const [busy, setBusy] = useState(false);
-  const [open, setOpen] = useState(null); // server detail
-
-  const load = useCallback(async () => {
-    try {
-      setData(await api("/mcp"));
-      setSpecs((await api("/mcp/specs")).specs);
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-  useEffect(() => {
-    load();
-  }, [load]);
-
-  const addSpec = useCallback(async () => {
-    if (!newSpec.trim()) return;
-    setBusy(true);
-    try {
-      const r = await api("/mcp/specs", { method: "POST", body: { spec: newSpec.trim() } });
-      setInfo(
-        r.requiresRestart ? "saved — restart `reasonix code` to bridge this server" : "saved",
-      );
-      setTimeout(() => setInfo(null), 4000);
-      setNewSpec("");
-      await load();
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setBusy(false);
-    }
-  }, [newSpec, load]);
-
-  const removeSpec = useCallback(
-    async (spec) => {
-      if (!confirm(`Remove MCP spec from config?\n\n${spec}`)) return;
-      setBusy(true);
-      try {
-        await api("/mcp/specs", { method: "DELETE", body: { spec } });
-        setInfo("removed — restart to drop the live bridge");
-        setTimeout(() => setInfo(null), 4000);
-        await load();
-      } catch (err) {
-        setError(err.message);
-      } finally {
-        setBusy(false);
-      }
-    },
-    [load],
-  );
-
-  if (!data && !error) return html`<div class="boot">loading MCP…</div>`;
-  if (error && !data) return html`<div class="notice err">${error}</div>`;
-
-  if (open) {
-    return html`
-      <div>
-        <div class="panel-header">
-          <h2 class="panel-title">MCP · ${open.label}</h2>
-          <button onClick=${() => setOpen(null)} style="margin-left: auto;">← back</button>
-        </div>
-        <div class="card">
-          <div class="row"><span class="card-title" style="margin: 0; flex: 0 0 110px;">spec</span><code>${open.spec}</code></div>
-          <div class="row"><span class="card-title" style="margin: 0; flex: 0 0 110px;">server</span><span>${open.serverInfo?.name ?? "—"} ${open.serverInfo?.version ? `v${open.serverInfo.version}` : ""}</span></div>
-          <div class="row"><span class="card-title" style="margin: 0; flex: 0 0 110px;">protocol</span><code>${open.protocolVersion}</code></div>
-        </div>
-        ${open.instructions ? html`<div class="notice">${open.instructions}</div>` : null}
-        <div class="section-title">Tools (${open.tools.length})</div>
-        <table>
-          <thead><tr><th>name</th><th>description</th></tr></thead>
-          <tbody>
-            ${open.tools.map((t) => html`<tr><td><code>${t.name}</code></td><td>${t.description ?? ""}</td></tr>`)}
-          </tbody>
-        </table>
-        ${
-          open.resources.length > 0
-            ? html`
-          <div class="section-title">Resources (${open.resources.length})</div>
-          <table>
-            <thead><tr><th>name</th><th>uri</th></tr></thead>
-            <tbody>
-              ${open.resources.map((r) => html`<tr><td>${r.name}</td><td><code>${r.uri}</code></td></tr>`)}
-            </tbody>
-          </table>
-        `
-            : null
-        }
-        ${
-          open.prompts.length > 0
-            ? html`
-          <div class="section-title">Prompts (${open.prompts.length})</div>
-          <table>
-            <thead><tr><th>name</th><th>description</th></tr></thead>
-            <tbody>
-              ${open.prompts.map((p) => html`<tr><td><code>${p.name}</code></td><td>${p.description ?? ""}</td></tr>`)}
-            </tbody>
-          </table>
-        `
-            : null
-        }
-      </div>
-    `;
-  }
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">MCP</h2>
-        <span class="panel-subtitle">${data.servers.length} bridged · ${specs?.length ?? 0} in config</span>
-      </div>
-      ${info ? html`<div class="notice">${info}</div>` : null}
-      ${error ? html`<div class="notice err">${error}</div>` : null}
-
-      <div class="section-title">Add server</div>
-      <div class="card row">
-        <input
-          type="text"
-          placeholder='spec — e.g. "fs=npx -y @modelcontextprotocol/server-filesystem /tmp/safe"'
-          value=${newSpec}
-          onInput=${(e) => setNewSpec(e.target.value)}
-        />
-        <button class="primary" disabled=${busy || !newSpec.trim()} onClick=${addSpec}>Add</button>
-      </div>
-
-      <div class="section-title">Bridged (${data.servers.length})</div>
-      ${
-        data.servers.length === 0
-          ? html`<div class="empty">No MCP servers in this session.</div>`
-          : html`
-          <table>
-            <thead><tr><th>label</th><th>spec</th><th class="numeric">tools</th><th></th></tr></thead>
-            <tbody>
-              ${data.servers.map(
-                (s) => html`
-                <tr key=${s.label} style="cursor: pointer;" onClick=${() => setOpen(s)}>
-                  <td><code>${s.label}</code></td>
-                  <td><code style="font-size: 11px;">${s.spec}</code></td>
-                  <td class="numeric">${fmtNum(s.toolCount)}</td>
-                  <td></td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-
-      <div class="section-title">Persisted specs (config.json)</div>
-      ${
-        (specs ?? []).length === 0
-          ? html`<div class="empty">No MCP specs persisted in <code>~/.reasonix/config.json</code>.</div>`
-          : html`
-          <table>
-            <thead><tr><th>spec</th><th></th></tr></thead>
-            <tbody>
-              ${specs.map(
-                (spec) => html`
-                <tr key=${spec}>
-                  <td><code>${spec}</code></td>
-                  <td class="numeric">
-                    <button class="danger" disabled=${busy} onClick=${(e) => {
-                      e.stopPropagation();
-                      removeSpec(spec);
-                    }}>remove</button>
-                  </td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-    </div>
-  `;
 }
 
 // ---------- shell ----------
